// bindings/src/types/Network.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Represents the available Tari networks. The variants and assigned byte needs to match the L1 network enum.
 */
export type Network = "mainnet" | "stagenet" | "nextnet" | "localnet" | "igor" | "esmeralda";